LayoutTests/fast/images/text-recognition/image-overlay-in-transparent-image-expected.txt

PASS boundingRect.width is 250
PASS boundingRect.height is 200
PASS successfullyParsed is true

TEST COMPLETE
